refactor: extract QuadTree from ForceDirectedLayout into own file
The Barnes-Hut QuadTree was a ~130-line static inner class inside
ForceDirectedLayout, making that file harder to navigate and preventing
reuse from other layout algorithms.

Changes:
- Extract QuadTree to gvisual/QuadTree.java as a package-private class
- Make applyRepulsion accept theta as a parameter instead of reading
  a constant from the enclosing class, improving configurability
- No behavioral changes; all existing call sites updated
